LldpNeighbor type
